% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussianMixtureModel.R
\name{gaussianMixtureModel}
\alias{gaussianMixtureModel}
\title{Find elements that are not in common}
\usage{
gaussianMixtureModel(expressionMatrix, genesOfInterest, k)
}
\arguments{
\item{expressionMatrix}{Numeric Matrix with genes as rows and samples as columns}

\item{genesOfInterest}{Character vector listing the genes you're interested in looking at}

\item{k}{Number of Gaussians to be used. This mixture model uses as many distributions as specified}
}
\value{
List with posterior probabilitiy data frames
}
\description{
Gaussian Mixture Model for gene expression distributions
}
\author{
Felipe Flores
}
