% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneExpressionHeatmap.R
\name{geneExpressionHeatmap}
\alias{geneExpressionHeatmap}
\title{Creating Gene Expression Heatmaps}
\usage{
geneExpressionHeatmap(expressionMatrix, genesOfInterest, samples,
  annotationDataFrame, clusterColumns = T, clusterRows = T, fontSize,
  kMeans = 1, title = character(0), annotationArguments = NULL, filename,
  palettes = c("Set1", "Dark2", "Set2", "Set3", "Accent", "Paired", "Pastel1",
  "Pastel2"), ...)
}
\arguments{
\item{expressionMatrix}{Matrix with genes as rows and samples as columns}

\item{genesOfInterest}{Vector of genes to be put on a heatmap}

\item{samples}{(optional) Indexing vector if only want to visualize a subset of samples. Can also be the sample names}

\item{annotationDataFrame}{Data frame with categories (clusters, groups, etc) that can be used to annotate a heatmap. Rows should be sample names and each column represents an annotation}

\item{fontSize}{Global font size}

\item{kMeans}{(optional) Number of k-means used to cluster rows}

\item{title}{(optional) Heatmap Title}

\item{annotationArgumnets}{(optional) List with further annotation parameters if needed}
}
\value{
hm, a heatmap object
}
\description{
Beautiful heatmap creator
}
\author{
Felipe Flores
}
