% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacore_pathway_FDR_plot.R
\name{metacore_pathway_FDR_plot}
\alias{metacore_pathway_FDR_plot}
\title{Metacore Pathway Plotter}
\usage{
metacore_pathway_FDR_plot(metacore_files, titles, outputnames)
}
\arguments{
\item{metacore_files}{Files with pathways from metacore in xlsx format. They originally come in xls format but this won't work. Must convert by hand with excel}

\item{titles}{Vector of plot titles. Example: c("Plot 1", "Plot 2")}

\item{outputnames}{Vector of names passed to ggsave. Example: c("plot1.pdf", "plot2.pdf")}
}
\value{
Saves plots as pdfs
}
\description{
Metacore Pathway Plotter
}
