% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outersect.R
\name{outersect}
\alias{outersect}
\title{Find elements that are not in common}
\usage{
outersect(x, y)
}
\arguments{
\item{x}{Element 1}

\item{y}{Element 2}
}
\value{
Sorted vector of elements that are different
}
\description{
Reverse of intersect
}
\author{
Felipe Flores
}
