% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleCorrelation.R
\name{sampleCorrelation}
\alias{sampleCorrelation}
\title{Sample Correlation Plot}
\usage{
sampleCorrelation(normalizedCountsTable)
}
\arguments{
\item{normalizedCountsTable}{Numeric matrix with genes as rows and samples as columns. Ideally variance-stabilized and normalized by DESeq methods}
}
\value{
Complex heatmap object
}
\description{
Function that creates a correlation plot with circles colored and sized based on correlation on the lower triangle and correlation values on the upper triangle
}
\examples{

}
