% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplePCA.R
\name{samplePCA}
\alias{samplePCA}
\title{Sample PCA plot for RNASeq}
\usage{
samplePCA(expressionMatrix, numberOfPCs = 2)
}
\arguments{
\item{expressionMatrix}{Data frame with expression data from a read.csv() function}

\item{numberOfPCs}{Integer value for number of principal components}
}
\value{
List with two plots, variancePlot has cumulative and individual importance of each PC, and pcaGRid is a grid of PCA plots
}
\description{
PCA plots of samples for DE analysis
}
\author{
Felipe Flores
}
