% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tSNE_multipleParameters.R
\name{tSNE_multipleParameters}
\alias{tSNE_multipleParameters}
\title{Multiple-parameter tSNE Run}
\usage{
tSNE_multipleParameters(matrix, perplexities, thetas, iterations, clusterDF,
  filename = "./plots/dimension-reduction/tSNE.pdf")
}
\arguments{
\item{matrix}{Matrix on which the runs of tSNE will be performed. Should have samples as columns}

\item{perplexities}{Vector of the perplexities to attempt}

\item{thetas}{Vector of the thetas to try}

\item{iterations}{Max Number of iterations (defaults to 800)}

\item{clusterDF}{Cluster data frame. Rows should be sample names and cells should have the name of the cluster for each sample}
}
\description{
Automated tSNE runs through a set of parameters specified by the user
}
\author{
Felipe Flores
}
