% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vennDiagrams.R
\name{vennDiagrams}
\alias{vennDiagrams}
\title{Generic Venn Diagram}
\usage{
vennDiagrams(listOfCohorts, fileName = NULL, title = NULL)
}
\arguments{
\item{listOfCohorts}{Sets to compare as a named list. Names will be used to label the diagram}

\item{title}{Title of Diagram}

\item{filename}{To be saved}
}
\value{
Sorted vector of elements that are different
}
\description{
Quick and Dirty Venn Diagram
}
\author{
Felipe Flores
}
