% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplePCA.R
\name{samplePCA}
\alias{samplePCA}
\title{Sample PCA plot for RNASeq}
\usage{
samplePCA(expressionMatrix, numberOfPCs = 2, clusters = 1,
  autoClustering = FALSE, colorByCluster = FALSE)
}
\arguments{
\item{expressionMatrix}{Data frame with variance-stabilized expression data (in DESeq would be obtained by assay(vsd), where vsd is a DESeq object after VST has been applied)}

\item{numberOfPCs}{Integer value for number of principal components}

\item{clusters}{User has the option of choosing how many clusters they want (k Means)}

\item{autoClustering}{Boolean. If TRUE the function will determine the optimal number of k means}

\item{colorByCluster}{Whether to color plot by cluster}
}
\value{
List with three plots, variancePlot has cumulative and individual importance of each PC, pcaGRid is a grid of PCA plots, and clusteringPlot is is the plot generated by eclust on the first 2 PCs
}
\description{
PCA plots of samples for DE analysis
}
\author{
Felipe Flores
}
